// src/app/model/user.module.ts
export class User {
    id: number;
    firstName: string;
    lastName: string;
    username: string;
    email: string;
    birthdate: string;
    zip: string;
    city: string;
    phone: string;
    usertype: string;
    password: string;
    monday: any;
    tuesday: any;
    wednesday: any;
    friday: any;
    thursday: any;
}